WIP: Experimentations on a Checkstyle test
The idea is to generate a checkstyle object from a codestyle object and to call the checkstyle library directly in the tests to verify that everything works as expected.
For now the call to checkstyle is ok (it works based on the console command) but we have to clean it.
We also have to make it an assertion and provide the tools that will help to write tests easily.